Add --scale and  --dry-run flags for flexibility

+++ tools/profile-openacc.sh
@@ -2,45 +2,61 @@
 # Simple bash script for lazy profiling. 
 cd "$(dirname "$0")"
 
-#NAMELIST_SIZE=128
-NAMELIST_SIZE=512
-#NAMELIST_SIZE=1024
-#NAMELIST_SIZE=4096
-
-
+# Configuration params
+DEFAULT_SCALE=512
 TARGET_DIR="../openacc"
-TARGET_NAMELIST="namelist.$NAMELIST_SIZE.profile"
+WORKING_DIR="."
 
+# Constants
 NAMELIST="namelist"
 EXECUTABLE="nemolite2d.exe"
-WORKING_DIR="."
 
 # Get the date time this was initiated.
 DATETIME=$(date '+%Y%m%d-%H%M%S')
-TIMELINE_FILE="$NAMELIST_SIZE-$DATETIME-timeline.nvvp"
-METRICS_FILE="$NAMELIST_SIZE--metrics.nvvp"
 
 # help message fn.
 usage(){
-    echo "usage: $0 [-h] [-m|--metrics]]"
+    echo "usage: $0 [-h] [-s|--scale SCALE] [-m|--metrics] [--dry-run]]"
 }
 
-
+# function to run th eprofile.
 profile(){
-    metrics=$1
+    SCALE=$1
+    METRICS=$2
+    DRYRUN=$3
+
+    TARGET_NAMELIST="namelist.$SCALE.profile"
+    TIMELINE_FILE="$SCALE-$DATETIME-timeline.nvvp"
+    METRICS_FILE="$SCALE-metrics.nvvp"
     # Log a messsage.
-    echo "Profiling $TARGET_NAMELIST using $TARGET_DIR/$EXECUTABLE"
+    echo "Profiling $SCALE using $TARGET_DIR/$EXECUTABLE"
+
+    # Check the namelist file exists.
+    if [ ! -f "$TARGET_DIR/$TARGET_NAMELIST" ]; then
+        echo "Error: $TARGET_DIR/$TARGET_NAMELIST does not exist"
+        return
+    fi
 
     # Update the namelist file (symlink)
     rm -f "$WORKING_DIR/$NAMELIST"
     ln -s "$TARGET_DIR/$TARGET_NAMELIST" "$WORKING_DIR/$NAMELIST"
 
     # Capture a timeline
-    nvprof -f -o "$TARGET_DIR/$TIMELINE_FILE" "$TARGET_DIR/$EXECUTABLE"
+    timeline_command="nvprof -f -o $TARGET_DIR/$TIMELINE_FILE $TARGET_DIR/$EXECUTABLE"
+    if [ "$DRYRUN" = "0" ]; then
+        $timeline_command
+    else
+        echo "$timeline_command"
+    fi 
 
     # optionally capture full details
-    if [ "$metrics" = "1" ]; then
-        nvprof -f -o "$TARGET_DIR/$METRICS_FILE" --analysis-metrics "$TARGET_DIR/$EXECUTABLE"
+    metrics_command="nvprof -f -o $TARGET_DIR/$METRICS_FILE --analysis-metrics $TARGET_DIR/$EXECUTABLE"
+    if [ "$METRICS" = "1" ]; then
+        if [ "$DRYRUN" = "0" ]; then
+            $metrics_command
+        else
+            echo "$metrics_command"
+        fi 
     fi
 
     # Revert the namelist file.
@@ -48,11 +64,18 @@ profile(){
 }
 
 # Check some usage and run the sript. 
+scale=$DEFAULT_SCALE
 metrics=0
+dryrun=0
 
 while [ "$1" != "" ]; do
     case $1 in
+        -s | --scale )          shift
+                                scale=$1
+                                ;;
         -m | --metrics )        metrics=1
+                                ;;
+             --dry-run )        dryrun=1
                                 ;;
         -h | --help )           usage
                                 exit
@@ -64,4 +87,4 @@ while [ "$1" != "" ]; do
 done
 
 # Run the profile fn
-profile $metrics
+profile $scale $metrics $dryrun
